A gameplay event system which supports generic and multiple parameters.
It can make code more clear to read as well as avoid the coupling of the code with other system.
To use this event dispatcher, declare the event of the game in GameEvents.cs(with clear and specific naming space), then simply listen the event and trigger it can execute the callback function.