A bracelet accessory located at player's wrist to show player's current hp.
Finish bracelet's art resource, including model, shading and material.
A script for bracelet to change its outlook once player's health change.
A accessory attachment for hand model. While the hand is instantiate during runtime, instead of adding the accessory directly to the hand's prefab (which will result in that the accessory not correctly created), using the script to attach those accessory for hands at specify bone, so that once the hands are instantiate or destroy at anytime of game, the accessory can follow the hands' life cycle.